Refactor folder structure in 'pages'
