+++ src/components/pages/health/Health.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from "react";
-import logo from "../../assets/images/logos/logo.png";
+import logo from "../../../assets/images/logos/logo.png";
 import {
   Button,
   Table,
@@ -11,14 +11,14 @@ import {
   callBackendHealth,
   DataIntegrity,
   SystemHealthData,
-} from "../../background/api/api";
-import { audioOnOff, setAudioVolumeByID } from "../../background/methods/sound";
-import { logout } from "../../background/api/auth";
-import { getDurationAsString } from "../../background/methods/time";
-import { hasKey } from "../../background/methods/ObjectKeysTS";
-import { formatBytes } from "../../background/methods/bytes";
-import { FFLoading } from "../../components/basicElements/Loading";
-import traffic_light from "../../assets/images/icons/material.io/traffic_light.svg";
+} from "../../../background/api/api";
+import { audioOnOff, setAudioVolumeByID } from "../../../background/methods/sound";
+import { logout } from "../../../background/api/auth";
+import { getDurationAsString } from "../../../background/methods/time";
+import { hasKey } from "../../../background/methods/ObjectKeysTS";
+import { formatBytes } from "../../../background/methods/bytes";
+import { FFLoading } from "../../basicElements/Loading";
+import traffic_light from "../../../assets/images/icons/material.io/traffic_light.svg";
 
 export default function Health() {
   const [systemHealthData, setSystemHealthData] = useState<
